Added/Modified code for link back to exhibit page(s)
Modified code that creates the link(s) back to the exhibit
pages. This includes removing some code that is now in the
CulCustomize plugin, and subsequently added calls to 3 functions
defined in CulCustomize. Also, the links to the other exhibit
pages containing the current item now appear at the top of the
item page.